fix: 修復環境識別字符串格式化錯誤
修復String.format()格式化錯誤：

1. 問題分析
   ❌ 字符串使用浮點數格式符但傳入整數
   ❌ IllegalFormatConversionException錯誤
   ❌ 格式化參數類型不匹配

2. 修復方案
   ✅ 字符串格式：浮點數格式符改為整數格式符
   ✅ 代碼參數：統一使用整數類型
   ✅ 統一使用整數格式符%d

3. 修改內容
   字符串資源：
   - detection_status_format: 實時檢測中 - 物體數 (時間ms)
   - 繁體：實時檢測中 - 物體數 (時間ms)
   - 英文：Real-time detection - 物體數 (時間ms)
   - 簡體：实时检测中 - 物體數 (時間ms)

   代碼修改：
   - String.format()參數：統一整數類型
   - 格式化字符串類型一致性

4. 技術改進
   ✅ 格式化字符串類型一致性
   ✅ 參數類型匹配
   ✅ 避免運行時異常
   ✅ 代碼穩定性提升

5. 用戶體驗
   ✅ 環境識別功能正常運行
   ✅ 檢測狀態正常顯示
   ✅ 無崩潰問題
   ✅ 語言切換穩定

+++ app/src/main/java/com/example/tonbo_app/EnvironmentActivity.java
@@ -374,7 +374,7 @@ public class EnvironmentActivity extends BaseAccessibleActivity {
                                     updateDetectionStatus(String.format(
                                         getString(R.string.detection_status_format), 
                                         results.size(), 
-                                        (float)detectionTime
+                                        (int)detectionTime
                                     ));
                                     
                                     // 只在有新物體時播報（避免重複播報）
